Replace old error handler
