+++ tests/error_handlers/test_sentry.py
@@ -9,21 +9,28 @@
 # Copyright (c) 2011 globo.com thumbor@googlegroups.com
 
 from preggy import expect
+from sentry_sdk import Hub
 
 from thumbor import __version__
-from thumbor.error_handlers.sentry import ErrorHandler
 from thumbor.context import ServerParameters
 from thumbor.config import Config
+from thumbor.server import get_importer
 
 from tests.base import TestCase
+from sentry_sdk.transport import Transport
 
 
-class FakeSentry(object):
-    def __init__(self, dsn):
-        self.captured_exceptions = []
+class MockTransport(Transport):
+    """
+    Transport that stores events in memory instead of sending them to Sentry.
+    """
 
-    def captureException(self, exception, *args, **kw):
-        self.captured_exceptions.append((exception, args, kw))
+    def __init__(self, options=None):
+        super(MockTransport, self).__init__(options=options)
+        self.captured_events = []
+
+    def capture_event(self, event):
+        self.captured_events.append(event)
 
 
 class FakeRequest(object):
@@ -50,15 +57,32 @@ class FakeHandler(object):
 
 
 class InvalidSentryTestCase(TestCase):
-    def test_when_invalid_configuration(self):
+    def test_when_missing_dsn(self):
         with expect.error_to_happen(RuntimeError):
-            ErrorHandler(self.config)
+            get_importer(Config(
+                SECURITY_KEY='ACME-SEC',
+                USE_CUSTOM_ERROR_HANDLING=True,
+                ERROR_HANDLER_MODULE='thumbor.error_handlers.sentry',
+            ))
+
+    def test_when_invalid_path(self):
+        with expect.error_to_happen(ImportError):
+            get_importer(Config(
+                SECURITY_KEY='ACME-SEC',
+                USE_CUSTOM_ERROR_HANDLING=True,
+                SENTRY_DSN_URL='https://key@sentry.io/12345',
+                ERROR_HANDLER_MODULE='thumbor.error_handlers.invalid',
+            ))
 
 
 class SentryTestCase(TestCase):
     def setUp(self, *args, **kwargs):
         super(SentryTestCase, self).setUp(*args, **kwargs)
-        self.client_mock = FakeSentry("FAKE DSN")
+        self.app = self.get_app()
+
+        self.transport_mock = MockTransport()
+        Hub.current.client.transport = self.transport_mock
+
         self.http_handler = FakeHandler()
 
     def get_server(self):
@@ -69,45 +93,38 @@ class SentryTestCase(TestCase):
     def get_config(self):
         return Config(
             SECURITY_KEY='ACME-SEC',
-            SENTRY_DSN_URL="http://sentry-dsn-url"
+            SENTRY_DSN_URL='https://key@sentry.io/12345',
+            USE_CUSTOM_ERROR_HANDLING=True,
+            ERROR_HANDLER_MODULE='thumbor.error_handlers.sentry',
         )
 
+    def get_importer(self):
+        return get_importer(self.config)
+
     def test_when_error_occurs_should_have_called_client(self):
-        handler = ErrorHandler(self.config, client=self.client_mock)
-        handler.handle_error(self.context, self.http_handler, RuntimeError("Test"))
+        self.importer.error_handler.handle_error(self.context, self.http_handler, RuntimeError("Test"))
 
-        expect(self.client_mock.captured_exceptions).not_to_be_empty()
-        expect(self.client_mock.captured_exceptions).to_length(1)
+        expect(self.transport_mock.captured_events).not_to_be_empty()
+        expect(self.transport_mock.captured_events).to_length(1)
 
-        exception, args, kw = self.client_mock.captured_exceptions[0]
-        expect(exception.__class__.__name__).to_equal("RuntimeError")
-        expect(kw).to_include('data')
-        expect(kw).to_include('extra')
+        event = self.transport_mock.captured_events[0]
 
-        data, extra = kw['data'], kw['extra']
+        expect(event['exception']['values'][0]['type']).to_equal("RuntimeError")
+
+        expect(event).to_include('extra')
+        expect(event).to_include('request')
+        request, extra = event['request'], event['extra']
 
         expect(extra).to_include('thumbor-version')
-        expect(extra['thumbor-version']).to_equal(__version__)
+        expect(extra['thumbor-version']).to_equal("'%s'" % __version__)
 
-        expect(extra).to_include('Headers')
-        expect(extra['Headers']).to_length(2)
+        expect(request).to_include('headers')
+        expect(request['headers']).to_length(2)
 
-        expect(extra['Headers']).to_include('Cookie')
-        expect(extra['Headers']['Cookie']).to_length(2)
+        expect(request['headers']).to_include('Cookie')
+        expect(request['headers']['Cookie']).to_equal('cookie1=value; cookie2=value2;')
+        expect(request['headers']['header1']).to_equal('value1')
 
-        expect(data['modules']).not_to_be_empty()
+        expect(request['query_string']).to_equal('a=1&b=2')
 
-        del data['modules']
-
-        expect(data).to_be_like({
-            'sentry.interfaces.Http': {
-                'url': "http://test/test/",
-                'method': "GET",
-                'data': [],
-                'body': "body",
-                'query_string': "a=1&b=2"
-            },
-            'sentry.interfaces.User': {
-                'ip': "127.0.0.1",
-            }
-        })
+        expect(event['modules']).not_to_be_empty()
